Revised NewTaxonomy
So far, the method NewTaxonomy took no arguments and returned the
complete taxonomy represented in the taxonomy database. Now NewTaxonomy
takes as argument a slice of taxon IDs. If this slice is nil or empty,
the old behavior is restored, the full taxonomy is returned. Otherwise,
we get a partial taxonomy with the supplied taxa as leaves.