Remove hyphens from list of genome accessions II
I thought I had fixed this on Apr 4. However, in that fix I had assumed
that hyphens represent empty query results and hence can only occupy the
first position in an accession list. As of the latest version of the
database, hyphens can appear anywhere in a list of genome accessions. So
now I search the full accessions list of each target and neighbor and
remove any hyphens.