New matching in fintac -t & -n
Matching of targets and neighbors with -t and -n was predicated on
prefixes. This has now been generalized to any regex. The default target
regex is "^t", and anything not matching this is a neighbor, unless a
neighbor regex is explicitly set. This should leave the default behavior
unchanged but allows filtering for other strings than common prefixes.