Improved fintac
So far, when fintac encountered a leaf label that was neither a target
nor a neighbor it aborted. Now it just gives a warning but carries out
the rest of the analysis. This allows users to include "unknown" genomes
in the phylogeny, which are then effectively ignored by fintac. The
tutorial shows an example where this is useful.