Changed error handling in tdb
All functions used to call util.Check(err) to check errors. This
function calls log.Fatal() if the error isn't nil, which is ok for a
command line program, but not for a web server. So now all functions
that may be called repeatedly from a web server return an error instead
of aborting upon error.