Now also consider merged nodes
So far, merged nodes were ignored, which could lead to genomes with
taxon IDs that do not appear in the taxon table. This is problematic on
two levels: First, genome.taxid is a foreign key referencing
taxon.taxid, but this key is currently not enforced. Second,
genome.taxids should simply be correct. This is now ok, genome.taxids
are now correct due to the inclusion of merged.dmp in the database
construction. However, the foreign key issue still awaits resolution.